ml_api/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
python-dotenv
requests
